Modernize the codebase
- Move `lib` to `src` and modularize
- Use ES exports
- Use `babel-tap` to run tests
- Create `lib` using Babel
- Create `dist` build using rollup
- Swap jshint for eslint
