+++ examples/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset='utf-8'>
     <title>Browser Bunyan Examples</title>
-    <script src="../dist/browser-bunyan.js"></script>
+    <script src="../dist/browser-bunyan.min.js"></script>
 </head>
 <body>
     <h1>Browser Bunyan examples</h1>
