--- lib/bunyan.js
@@ -1,967 +1,0 @@
-/**
- * !This is a stripped down version of Bunyan targeted specifically for the browser
- *
- * -------------------------------------------------------------------------------
- *
- * Copyright (c) 2014 Trent Mick. All rights reserved.
- * Copyright (c) 2014 Joyent Inc. All rights reserved.
- *
- * The bunyan logging library for node.js.
- *
- * -*- mode: js -*-
- * vim: expandtab:ts=4:sw=4
- */
-    
-'use strict';
-
-var VERSION = '0.4.0';
-
-// Bunyan log format version. This becomes the 'v' field on all log records.
-// `0` is until I release a version '1.0.0' of node-bunyan. Thereafter,
-// starting with `1`, this will be incremented if there is any backward
-// incompatible change to the log record format. Details will be in
-// 'CHANGES.md' (the change log).
-var LOG_VERSION = 0;
-
-//---- Internal support stuff
-
-var CALL_STACK_ERROR = 'call-stack-error';
-
-/**
- * A shallow copy of an object. Bunyan logging attempts to never cause
- * exceptions, so this function attempts to handle non-objects gracefully.
- */
-function objCopy(obj) {
-    if (typeof obj === 'undefined' || obj === null) {  // null or undefined
-        return obj;
-    } else if (Array.isArray(obj)) {
-        return obj.slice();
-    } else if (typeof (obj) === 'object') {
-        var copy = {};
-        Object.keys(obj).forEach(function (k) {
-            copy[k] = obj[k];
-        });
-        return copy;
-    } else {
-        return obj;
-    }
-}
-
-//---- These are simplified versions of util.format without importing the whole module, which would be bulky when browserified
-
-var inspect = function(obj) {
-    if(typeof obj === 'undefined') {
-        return 'undefined';
-    }
-    if(obj === null) {
-        return 'null';
-    }
-    if(Array.isArray(obj)) {
-        var items = obj.map(function(obj) {
-            return inspect(obj);
-        });
-        return '[ ' + items.join(', ') + ' ]';
-    }
-    if(typeof obj === 'object') {
-        return JSON.stringify(obj);
-    }
-    if(typeof obj === 'function') {
-        return '[Function: ' + obj.name + ']';
-    }
-    if(typeof obj === 'boolean' || typeof obj === 'number') {
-        return obj;
-    }
-    return '\'' + obj.toString() + '\'';
-};
-
-var format = function(f) {
-
-    if (typeof f !== 'string') {
-        var objects = new Array(arguments.length);
-        for (var index = 0; index < arguments.length; index++) {
-            objects[index] = inspect(arguments[index]);
-        }
-        return objects.join(' ');
-    }
-
-
-    var formatRegExp = /%[sdj%]/g;
-
-    var i = 1;
-    var args = arguments;
-    var len = args.length;
-    var str = String(f).replace(formatRegExp, function(x) {
-        if (x === '%%') {
-            return '%';
-        }
-        if (i >= len) {
-            return x;
-        }
-        switch (x) {
-            case '%s': return String(args[i++]);
-            case '%d': return Number(args[i++]);
-            case '%j':
-                try {
-                    return JSON.stringify(args[i++]);
-                } catch (_) {
-                    return '[Circular]';
-                }
-                break;
-            default:
-                return x;
-        }
-    });
-    for (var x = args[i]; i < len; x = args[++i]) {
-        str += ' ' + x;
-    }
-    return str;
-};
-
-function extractSrcFromStacktrace(stack, level) {
-
-    var stackLines = stack.split('\n');
-
-    //chrome starts with error
-    if(stackLines[0] && stackLines[0].indexOf(CALL_STACK_ERROR) >= 0) {
-        stackLines.shift();
-    }
-
-    //the line of the stacktrace
-    var targetLine = stackLines[level];
-    var lineInfo = null;
-    if(targetLine) {
-        var execResult = /^\s*(at|.*@)\s*(.+)?$/.exec(targetLine);
-        if(Array.isArray(execResult) && execResult[2]) {
-            lineInfo = execResult[2];
-        } else {
-            lineInfo = targetLine;    
-        }
-    }
-    return lineInfo;
-}
-
-function _indent(s, indent) {
-    if (!indent) {
-        indent = '    ';
-    }
-    var lines = s.split(/\r?\n/g);
-    return indent + lines.join('\n' + indent);
-}
-
-
-/**
- * Warn about an bunyan processing error.
- *
- * @param msg {String} Message with which to warn.
- * @param dedupKey {String} Optional. A short string key for this warning to
- *      have its warning only printed once.
- */
-function _warn(msg, dedupKey) {
-    if (dedupKey) {
-        if (_warned[dedupKey]) {
-            return;
-        }
-        _warned[dedupKey] = true;
-    }
-    console.error(msg + '\n');
-}
-function _haveWarned(dedupKey) {
-    return _warned[dedupKey];
-}
-var _warned = {};
-
-
-function ConsoleRawStream() {
-}
-ConsoleRawStream.prototype.write = function (rec) {
-    if (rec.level < INFO) {
-        console.log(rec);
-    } else if (rec.level < WARN) {
-        console.info(rec);
-    } else if (rec.level < ERROR) {
-        console.warn(rec);
-    } else {
-        console.error(rec);
-    }
-
-    if(rec.err && rec.err.stack) {
-        console.error(rec.err.stack);
-    }
-    if(rec.obj) {
-        console.log(rec.obj);
-    }
-};
-
-function ConsoleFormattedStream(opts) {
-    opts = opts || {};
-    this.logByLevel = !!opts.logByLevel;
-}
-ConsoleFormattedStream.prototype.write = function (rec) {
-
-    var levelCss,
-        defaultCss = 'color: DimGray',
-        msgCss = 'color: SteelBlue',
-        srcCss = 'color: DimGray; font-style: italic; font-size: 0.9em',
-        consoleMethod;
-
-    var loggerName = rec.childName ? rec.name + '/' + rec.childName : rec.name;
-
-    //get level name and pad start with spacs
-    var levelName = nameFromLevel[rec.level];
-    var formattedLevelName = (Array(6 - levelName.length).join(' ') + levelName).toUpperCase();
-
-    if(this.logByLevel) {
-        if(rec.level === TRACE) {
-            levelName = 'debug';
-        } else if(rec.level === FATAL) {
-            levelName = 'error';
-        }
-        consoleMethod = typeof console[levelName] === 'function' ? console[levelName] : console.log;
-    } else {
-        consoleMethod = console.log;
-    }
-
-    if (rec.level < DEBUG) {
-        levelCss = 'color: DeepPink';
-    } else if (rec.level < INFO) {
-        levelCss = 'color: GoldenRod';
-    } else if (rec.level < WARN) {
-        levelCss = 'color: DarkTurquoise';
-    } else if (rec.level < ERROR) {
-        levelCss = 'color: Purple';
-    } else if (rec.level < FATAL) {
-        levelCss = 'color: Crimson';
-    } else {
-        levelCss = 'color: Black';
-    }
-
-    function padZeros(number, len) {
-        return Array((len + 1) - (number + '').length).join('0') + number;
-    }
-
-    var logArgs = [];
-    logArgs.push('[%s:%s:%s:%s] %c%s%c: %s: %c%s %c%s');
-    logArgs.push(padZeros(rec.time.getHours(), 2));
-    logArgs.push(padZeros(rec.time.getMinutes(), 2));
-    logArgs.push(padZeros(rec.time.getSeconds(), 2));
-    logArgs.push(padZeros(rec.time.getMilliseconds(), 4));
-    logArgs.push(levelCss);
-    logArgs.push(formattedLevelName);
-    logArgs.push(defaultCss);
-    logArgs.push(loggerName);
-    logArgs.push(msgCss);
-    logArgs.push(rec.msg);
-    if(rec.src) {
-        logArgs.push(srcCss);
-        logArgs.push(rec.src);
-    }
-
-    consoleMethod.apply(console, logArgs);
-    if(rec.err && rec.err.stack) {
-        consoleMethod.call(console, '%c%s,', levelCss, rec.err.stack);
-    }
-    if(rec.obj) {
-        consoleMethod.call(console, rec.obj);
-    }
-};
-
-//---- Levels
-
-var TRACE = 10;
-var DEBUG = 20;
-var INFO = 30;
-var WARN = 40;
-var ERROR = 50;
-var FATAL = 60;
-
-var levelFromName = {
-    'trace': TRACE,
-    'debug': DEBUG,
-    'info': INFO,
-    'warn': WARN,
-    'error': ERROR,
-    'fatal': FATAL
-};
-var nameFromLevel = {};
-Object.keys(levelFromName).forEach(function (name) {
-    nameFromLevel[levelFromName[name]] = name;
-});
-
-/**
- * Resolve a level number, name (upper or lowercase) to a level number value.
- *
- * @api public
- */
-function resolveLevel(nameOrNum) {
-    var level = (typeof (nameOrNum) === 'string' ? levelFromName[nameOrNum.toLowerCase()] : nameOrNum);
-    return level;
-}
-
-//---- Logger class
-
-/**
- * Create a Logger instance.
- *
- * @param options {Object} See documentation for full details. At minimum
- *    this must include a 'name' string key. Configuration keys:
- *      - `streams`: specify the logger output streams. This is an array of
- *        objects with these fields:
- *          - `type`: The stream type. See README.md for full details.
- *            Often this is implied by the other fields. Examples are
- *            'file', 'stream' and "raw".
- *          - `level`: Defaults to 'info'.
- *          - `path` or `stream`: The specify the file path or writeable
- *            stream to which log records are written. E.g.
- *            `stream: process.stdout`.
- *          - `closeOnExit` (boolean): Optional. Default is true for a
- *            'file' stream when `path` is given, false otherwise.
- *        See README.md for full details.
- *      - `level`: set the level for a single output stream (cannot be used
- *        with `streams`)
- *      - `stream`: the output stream for a logger with just one, e.g.
- *        `process.stdout` (cannot be used with `streams`)
- *      - `serializers`: object mapping log record field names to
- *        serializing functions. See README.md for details.
- *      - `src`: Boolean (default false). Set true to enable 'src' automatic
- *        field with log call source info.
- *    All other keys are log record fields.
- *
- * An alternative *internal* call signature is used for creating a child:
- *    new Logger(<parent logger>, <child options>[, <child opts are simple>]);
- *
- * @param _childSimple (Boolean) An assertion that the given `_childOptions`
- *    (a) only add fields (no config) and (b) no serialization handling is
- *    required for them. IOW, this is a fast path for frequent child
- *    creation.
- */
-function Logger(options, _childOptions, _childSimple) {
-    if (!(this instanceof Logger)) {
-        return new Logger(options, _childOptions);
-    }
-
-    // Input arg validation.
-    var parent;
-    if (_childOptions !== undefined) {
-        parent = options;
-        options = _childOptions;
-        if (!(parent instanceof Logger)) {
-            throw new TypeError(
-                'invalid Logger creation: do not pass a second arg');
-        }
-    }
-    if (!options) {
-        throw new TypeError('options (object) is required');
-    }
-    if (!parent) {
-        if (!options.name) {
-            throw new TypeError('options.name (string) is required');
-        }
-    } else {
-        if (options.name) {
-            throw new TypeError(
-                'invalid options.name: child cannot set logger name');
-        }
-    }
-    if (options.stream && options.streams) {
-        throw new TypeError('cannot mix "streams" and "stream" options');
-    }
-    if (options.streams && !Array.isArray(options.streams)) {
-        throw new TypeError('invalid options.streams: must be an array');
-    }
-    if (options.serializers && (typeof (options.serializers) !== 'object' || Array.isArray(options.serializers))) {
-        throw new TypeError('invalid options.serializers: must be an object');
-    }
-
-    var fields, name, i;
-
-    // Fast path for simple child creation.
-    if (parent && _childSimple) {
-        // `_isSimpleChild` is a signal to stream close handling that this child
-        // owns none of its streams.
-        this._isSimpleChild = true;
-
-        this._level = parent._level;
-        this.streams = parent.streams;
-        this.serializers = parent.serializers;
-        this.src = parent.src;
-        fields = this.fields = {};
-        var parentFieldNames = Object.keys(parent.fields);
-        for (i = 0; i < parentFieldNames.length; i++) {
-            name = parentFieldNames[i];
-            fields[name] = parent.fields[name];
-        }
-        var names = Object.keys(options);
-        for (i = 0; i < names.length; i++) {
-            name = names[i];
-            fields[name] = options[name];
-        }
-        return;
-    }
-
-    // Null values.
-    var self = this;
-    if (parent) {
-        this._level = parent._level;
-        this.streams = [];
-        for (i = 0; i < parent.streams.length; i++) {
-            var s = objCopy(parent.streams[i]);
-            s.closeOnExit = false; // Don't own parent stream.
-            this.streams.push(s);
-        }
-        this.serializers = objCopy(parent.serializers);
-        this.src = parent.src;
-        this.fields = objCopy(parent.fields);
-        if (options.level) {
-            this.level(options.level);
-        }
-    } else {
-        this._level = Number.POSITIVE_INFINITY;
-        this.streams = [];
-        this.serializers = null;
-        this.src = false;
-        this.fields = {};
-    }
-
-    // Handle *config* options (i.e. options that are not just plain data
-    // for log records).
-    if (options.stream) {
-        self.addStream({
-            type: 'stream',
-            stream: options.stream,
-            closeOnExit: false,
-            level: options.level
-        });
-    } else if (options.streams) {
-        options.streams.forEach(function (s) {
-            self.addStream(s, options.level);
-        });
-    } else if (parent && options.level) {
-        this.level(options.level);
-    } else if (!parent) {
-
-        /*
-         * In the browser we'll be emitting to console.log by default.
-         * Any console.log worth its salt these days can nicely render
-         * and introspect objects (e.g. the Firefox and Chrome console)
-         * so let's emit the raw log record. Are there browsers for which
-         * that breaks things?
-         */
-        self.addStream({
-            type: 'raw',
-            stream: new ConsoleRawStream(),
-            closeOnExit: false,
-            level: options.level
-        });
-
-    }
-    if (options.serializers) {
-        self.addSerializers(options.serializers);
-    }
-    if (options.src) {
-        this.src = true;
-    }
-
-    // Fields.
-    // These are the default fields for log records (minus the attributes
-    // removed in this constructor). To allow storing raw log records
-    // (unrendered), `this.fields` must never be mutated. Create a copy for
-    // any changes.
-    fields = objCopy(options);
-    delete fields.stream;
-    delete fields.level;
-    delete fields.streams;
-    delete fields.serializers;
-    delete fields.src;
-    if (this.serializers) {
-        this._applySerializers(fields);
-    }
-    Object.keys(fields).forEach(function (k) {
-        self.fields[k] = fields[k];
-    });
-}
-
-/**
- * Add a stream
- *
- * @param stream {Object}. Object with these fields:
- *    - `type`: The stream type. See README.md for full details.
- *      Often this is implied by the other fields. Examples are
- *      'file', 'stream' and "raw".
- *    - `path` or `stream`: The specify the file path or writeable
- *      stream to which log records are written. E.g.
- *      `stream: process.stdout`.
- *    - `level`: Optional. Falls back to `defaultLevel`.
- *    - `closeOnExit` (boolean): Optional. Default is true for a
- *      'file' stream when `path` is given, false otherwise.
- *    See README.md for full details.
- * @param defaultLevel {Number|String} Optional. A level to use if
- *      `stream.level` is not set. If neither is given, this defaults to INFO.
- */
-Logger.prototype.addStream = function addStream(s, defaultLevel) {
-    var self = this;
-    if (defaultLevel === null || defaultLevel === undefined) {
-        defaultLevel = INFO;
-    }
-
-    s = objCopy(s);
-
-    //in browser bunyan, streams are always raw
-    s.type = 'raw';
-
-    if (s.level) {
-        s.level = resolveLevel(s.level);
-    } else {
-        s.level = resolveLevel(defaultLevel);
-    }
-    if (s.level < self._level) {
-        self._level = s.level;
-    }
-
-    if (!s.closeOnExit) {
-        s.closeOnExit = false;
-    }
-
-    self.streams.push(s);
-    delete self.haveNonRawStreams;  // reset
-};
-
-
-/**
- * Add serializers
- *
- * @param serializers {Object} Optional. Object mapping log record field names
- *    to serializing functions. See README.md for details.
- */
-Logger.prototype.addSerializers = function addSerializers(serializers) {
-    var self = this;
-
-    if (!self.serializers) {
-        self.serializers = {};
-    }
-    Object.keys(serializers).forEach(function (field) {
-        var serializer = serializers[field];
-        if (typeof (serializer) !== 'function') {
-            throw new TypeError(format(
-                'invalid serializer for "%s" field: must be a function',
-                field));
-        } else {
-            self.serializers[field] = serializer;
-        }
-    });
-};
-
-
-/**
- * Create a child logger, typically to add a few log record fields.
- *
- * This can be useful when passing a logger to a sub-component, e.g. a
- * 'wuzzle' component of your service:
- *
- *    var wuzzleLog = log.child({component: 'wuzzle'})
- *    var wuzzle = new Wuzzle({..., log: wuzzleLog})
- *
- * Then log records from the wuzzle code will have the same structure as
- * the app log, *plus the component='wuzzle' field*.
- *
- * @param options {Object} Optional. Set of options to apply to the child.
- *    All of the same options for a new Logger apply here. Notes:
- *      - The parent's streams are inherited and cannot be removed in this
- *        call. Any given `streams` are *added* to the set inherited from
- *        the parent.
- *      - The parent's serializers are inherited, though can effectively be
- *        overwritten by using duplicate keys.
- *      - Can use `level` to set the level of the streams inherited from
- *        the parent. The level for the parent is NOT affected.
- * @param simple {Boolean} Optional. Set to true to assert that `options`
- *    (a) only add fields (no config) and (b) no serialization handling is
- *    required for them. IOW, this is a fast path for frequent child
- *    creation. See 'tools/timechild.js' for numbers.
- */
-Logger.prototype.child = function (options, simple) {
-    return new (this.constructor)(this, options || {}, simple);
-};
-
-/**
- * Get/set the level of all streams on this logger.
- *
- * Get Usage:
- *    // Returns the current log level (lowest level of all its streams).
- *    log.level() -> INFO
- *
- * Set Usage:
- *    log.level(INFO)       // set all streams to level INFO
- *    log.level('info')     // can use 'info' et al aliases
- */
-Logger.prototype.level = function level(value) {
-    if (value === undefined) {
-        return this._level;
-    }
-    var newLevel = resolveLevel(value);
-    var len = this.streams.length;
-    for (var i = 0; i < len; i++) {
-        this.streams[i].level = newLevel;
-    }
-    this._level = newLevel;
-};
-
-
-/**
- * Get/set the level of a particular stream on this logger.
- *
- * Get Usage:
- *    // Returns an array of the levels of each stream.
- *    log.levels() -> [TRACE, INFO]
- *
- *    // Returns a level of the identified stream.
- *    log.levels(0) -> TRACE      // level of stream at index 0
- *    log.levels('foo')           // level of stream with name 'foo'
- *
- * Set Usage:
- *    log.levels(0, INFO)         // set level of stream 0 to INFO
- *    log.levels(0, 'info')       // can use 'info' et al aliases
- *    log.levels('foo', WARN)     // set stream named 'foo' to WARN
- *
- * Stream names: When streams are defined, they can optionally be given
- * a name. For example,
- *       log = new Logger({
- *         streams: [
- *           {
- *             name: 'foo',
- *             path: '/var/log/my-service/foo.log'
- *             level: 'trace'
- *           },
- *         ...
- *
- * @param name {String|Number} The stream index or name.
- * @param value {Number|String} The level value (INFO) or alias ('info').
- *    If not given, this is a 'get' operation.
- * @throws {Error} If there is no stream with the given name.
- */
-Logger.prototype.levels = function levels(name, value) {
-    if (name === undefined) {
-        return this.streams.map(
-            function (s) {
-                return s.level;
-            });
-    }
-    var stream;
-    if (typeof (name) === 'number') {
-        stream = this.streams[name];
-        if (stream === undefined) {
-            throw new Error('invalid stream index: ' + name);
-        }
-    } else {
-        var len = this.streams.length;
-        for (var i = 0; i < len; i++) {
-            var s = this.streams[i];
-            if (s.name === name) {
-                stream = s;
-                break;
-            }
-        }
-        if (!stream) {
-            throw new Error(format('no stream with name "%s"', name));
-        }
-    }
-    if (value === undefined) {
-        return stream.level;
-    } else {
-        var newLevel = resolveLevel(value);
-        stream.level = newLevel;
-        if (newLevel < this._level) {
-            this._level = newLevel;
-        }
-    }
-};
-
-
-/**
- * Apply registered serializers to the appropriate keys in the given fields.
- *
- * Pre-condition: This is only called if there is at least one serializer.
- *
- * @param fields (Object) The log record fields.
- * @param excludeFields (Object) Optional mapping of keys to `true` for
- *    keys to NOT apply a serializer.
- */
-Logger.prototype._applySerializers = function (fields, excludeFields) {
-    var self = this;
-
-    // Check each serializer against these (presuming number of serializers
-    // is typically less than number of fields).
-    Object.keys(this.serializers).forEach(function (name) {
-        if (fields[name] === undefined ||
-            (excludeFields && excludeFields[name])) {
-            return;
-        }
-        try {
-            fields[name] = self.serializers[name](fields[name]);
-        } catch (err) {
-            _warn(format('bunyan: ERROR: Exception thrown from the "%s" ' +
-                    'Bunyan serializer. This should never happen. This is a bug' +
-                    'in that serializer function.\n%s',
-                name, err.stack || err));
-            fields[name] = format('(Error in Bunyan log "%s" serializer broke field. See stderr for details.)', name);
-        }
-    });
-};
-
-
-/**
- * Emit a log record.
- *
- * @param rec {log record}
- * @param noemit {Boolean} Optional. Set to true to skip emission
- *      and just return the JSON string.
- */
-Logger.prototype._emit = function (rec, noemit) {
-    var i;
-
-    // Lazily determine if this Logger has non-'raw' streams. If there are
-    // any, then we need to stringify the log record.
-    if (this.haveNonRawStreams === undefined) {
-        this.haveNonRawStreams = false;
-        for (i = 0; i < this.streams.length; i++) {
-            if (!this.streams[i].raw) {
-                this.haveNonRawStreams = true;
-                break;
-            }
-        }
-    }
-
-    // Stringify the object. Attempt to warn/recover on error.
-    var str;
-    if (noemit || this.haveNonRawStreams) {
-        try {
-            str = JSON.stringify(rec, safeCycles()) + '\n';
-        } catch (e) {
-            var dedupKey = e.stack.split(/\n/g, 2).join('\n');
-            _warn('bunyan: ERROR: Exception in ' +
-                    '`JSON.stringify(rec)`. You can install the ' +
-                    '"safe-json-stringify" module to have Bunyan fallback ' +
-                    'to safer stringification. Record:\n' +
-                    _indent(format('%s\n%s', rec, e.stack)),
-                dedupKey);
-            str = format('(Exception in JSON.stringify(rec): %j. See stderr for details.)\n', e.message);
-
-        }
-    }
-
-    if (noemit) {
-        return str;
-    }
-
-
-    var level = rec.level;
-    for (i = 0; i < this.streams.length; i++) {
-        var s = this.streams[i];
-        if (s.level <= level) {
-            s.stream.write(rec);
-        }
-    }
-
-    return str;
-};
-
-
-/**
- * Build a log emitter function for level minLevel. I.e. this is the
- * creator of `log.info`, `log.error`, etc.
- */
-function mkLogEmitter(minLevel) {
-    return function () {
-        var log = this;
-
-        function mkRecord(args) {
-            var excludeFields;
-            if (args[0] instanceof Error) {
-                // `log.<level>(err, ...)`
-                fields = {
-                    // Use this Logger's err serializer, if defined.
-                    err: (log.serializers && log.serializers.err ? log.serializers.err(args[0]) : Logger.stdSerializers.err(args[0]))
-                };
-                excludeFields = {err: true};
-                if (args.length === 1) {
-                    msgArgs = [fields.err.message];
-                } else {
-                    msgArgs = Array.prototype.slice.call(args, 1);
-                }
-            } else if (typeof (args[0]) !== 'object' && args[0] !== null || Array.isArray(args[0])) {
-                // `log.<level>(msg, ...)`
-                fields = null;
-                msgArgs = Array.prototype.slice.call(args);
-            } else {
-                // `log.<level>(fields, msg, ...)`
-                fields = args[0];
-                if (args.length === 1 && fields.err && fields.err instanceof Error)
-                {
-                    msgArgs = [fields.err.message];
-                } else {
-                    msgArgs = Array.prototype.slice.call(args, 1);
-                }
-            }
-
-            // Build up the record object.
-            var rec = objCopy(log.fields);
-            rec.level = minLevel;
-            var recFields = (fields ? objCopy(fields) : null);
-            if (recFields) {
-                if (log.serializers) {
-                    log._applySerializers(recFields, excludeFields);
-                }
-                Object.keys(recFields).forEach(function (k) {
-                    rec[k] = recFields[k];
-                });
-            }
-            rec.levelName = nameFromLevel[minLevel];
-            rec.msg = msgArgs.length ? format.apply(log, msgArgs) : '';
-            if (!rec.time) {
-                rec.time = (new Date());
-            }
-            // Get call source info
-            if (log.src && !rec.src) {
-                try {
-                    //need to throw the error so there is a stack in IE		 	
-                    throw new Error(CALL_STACK_ERROR);
-                } catch(err) {
-                    var src = extractSrcFromStacktrace(err.stack, 2);
-                    if(!src && !_haveWarned('src')) {
-                        _warn('Unable to determine src line info', 'src');    
-                    }
-                    rec.src = src || '';
-                }
-            }
-            rec.v = LOG_VERSION;
-            return rec;
-        }
-
-        var fields = null;
-        var msgArgs = arguments;
-        var rec = null;
-        if (arguments.length === 0) {   // `log.<level>()`
-            return (this._level <= minLevel);
-        } else if (this._level > minLevel) {
-            /* pass through */
-        } else {
-            rec = mkRecord(msgArgs);
-            this._emit(rec);
-        }
-    };
-}
-
-
-/**
- * The functions below log a record at a specific level.
- *
- * Usages:
- *    log.<level>()  -> boolean is-trace-enabled
- *    log.<level>(<Error> err, [<string> msg, ...])
- *    log.<level>(<string> msg, ...)
- *    log.<level>(<object> fields, <string> msg, ...)
- *
- * where <level> is the lowercase version of the log level. E.g.:
- *
- *    log.info()
- *
- * @params fields {Object} Optional set of additional fields to log.
- * @params msg {String} Log message. This can be followed by additional
- *    arguments that are handled like
- *    [util.format](http://nodejs.org/docs/latest/api/all.html#util.format).
- */
-Logger.prototype.trace = mkLogEmitter(TRACE);
-Logger.prototype.debug = mkLogEmitter(DEBUG);
-Logger.prototype.info = mkLogEmitter(INFO);
-Logger.prototype.warn = mkLogEmitter(WARN);
-Logger.prototype.error = mkLogEmitter(ERROR);
-Logger.prototype.fatal = mkLogEmitter(FATAL);
-
-
-//---- Standard serializers
-// A serializer is a function that serializes a JavaScript object to a
-// JSON representation for logging. There is a standard set of presumed
-// interesting objects in node.js-land.
-
-Logger.stdSerializers = {};
-
-/*
- * This function dumps long stack traces for exceptions having a cause()
- * method. The error classes from
- * [verror](https://github.com/davepacheco/node-verror) and
- * [restify v2.0](https://github.com/mcavage/node-restify) are examples.
- *
- * Based on `dumpException` in
- * https://github.com/davepacheco/node-extsprintf/blob/master/lib/extsprintf.js
- */
-function getFullErrorStack(ex) {
-    var ret = ex.stack || ex.toString();
-    if (ex.cause && typeof (ex.cause) === 'function') {
-        var cex = ex.cause();
-        if (cex) {
-            ret += '\nCaused by: ' + getFullErrorStack(cex);
-        }
-    }
-    return (ret);
-}
-
-// Serialize an Error object
-// (Core error properties are enumerable in node 0.4, not in 0.6).
-Logger.stdSerializers.err = function(err) {
-    if (!err || !err.stack) {
-        return err;
-    }
-
-    var obj = {
-        message: err.message,
-        name: err.name,
-        stack: getFullErrorStack(err),
-        code: err.code,
-        signal: err.signal
-    };
-    return obj;
-};
-
-// A JSON stringifier that handles cycles safely.
-// Usage: JSON.stringify(obj, safeCycles())
-function safeCycles() {
-    var seen = [];
-    return function (key, val) {
-        if (!val || typeof (val) !== 'object') {
-            return val;
-        }
-        if (seen.indexOf(val) !== -1) {
-            return '[Circular]';
-        }
-        seen.push(val);
-        return val;
-    };
-}
-
-//---- Exports
-
-module.exports = Logger;
-
-module.exports.TRACE = TRACE;
-module.exports.DEBUG = DEBUG;
-module.exports.INFO = INFO;
-module.exports.WARN = WARN;
-module.exports.ERROR = ERROR;
-module.exports.FATAL = FATAL;
-module.exports.resolveLevel = resolveLevel;
-module.exports.levelFromName = levelFromName;
-module.exports.nameFromLevel = nameFromLevel;
-
-module.exports.VERSION = VERSION;
-module.exports.LOG_VERSION = LOG_VERSION;
-
-module.exports.createLogger = function createLogger(options) {
-    return new Logger(options);
-};
-
-// Useful for custom `type == 'raw'` streams that may do JSON stringification
-// of log records themselves. Usage:
-//    var str = JSON.stringify(rec, bunyan.safeCycles());
-module.exports.safeCycles = safeCycles;
-
-//streams
-module.exports.ConsoleFormattedStream = ConsoleFormattedStream;
-module.exports.ConsoleRawStream = ConsoleRawStream;
