+++ test/raw-stream.test.js
@@ -4,16 +4,10 @@
  * Test `type: 'raw'` Logger streams.
  */
 
-var format = require('util').format;
-var Logger = require('../lib/bunyan');
+import { Logger } from '../src/index';
 
-// node-tap API
-if (require.cache[__dirname + '/tap4nodeunit.js'])
-        delete require.cache[__dirname + '/tap4nodeunit.js'];
-var tap4nodeunit = require('./tap4nodeunit.js');
-var after = tap4nodeunit.after;
-var before = tap4nodeunit.before;
-var test = tap4nodeunit.test;
+import { test, beforEach as before, afterEach as after } from "babel-tap";
+
 
 
 function CapturingStream(recs) {
